Move the transcoding information from a transcode_info array to the client_types array.
At first I was reluctant to do this, but the more I think about it, the more it makes sense.
The client-specific are in fact a property of a client, because they depend on the
client features and thus it makes sense to store them along other client type informations.